More test fixes
